requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
ultralytics
torch
albumentations
streamlit_webrtc
av